added OOP seed and database fixes
