+++ seeds/orders.js
@@ -1,12 +1,28 @@
-exports.seed = function(knex, Promise) {
-    // Deletes ALL existing entries
-    return knex('orders').del()
-        .then(function() {
-            // Inserts seed entries
-            return knex('orders').insert([
-                { id: 1, Name: 'Goliath Anderson', address: '21 Monggomer st', email: 'dark&stormy@longisland.com', delivery: false, status: '' },
-                { id: 2, colName: 'Peter GRiffin', address: '72 Grogg Lane', email: 'Groggy@bigd.com', delivery: false, status: '' },
-                { id: 3, colName: 'Nickel Sahma', address: '69 Virgina Beach', email: 'virgin@beach.com', delivery: false, status: '' }
-            ]);
-        });
-};
+const uuidv4 = require('uuid/v4')
+const faker = require('faker')
+const knex = require('knex')
+let A = []
+function createData () {
+  Array(3).fill(1).forEach(x => {
+    A.push(new Order((faker.name.firstName() + ' ' + faker.name.lastName()), (faker.address.city() + ' ' + faker.address.streetAddress()), faker.internet.email(), faker.random.boolean(), faker.random.arrayElement(['delivered','processing','ready'])))
+    console.log('beep')
+  })
+}
+function Order (name, address, email, delivery, status) {
+  this.name = name
+  this.address = address
+  this.created_at = Date.now()
+  this.email = email
+  this.delivery = delivery
+  this.status = status
+}
+
+exports.seed = function (knex, Promise) {
+  // Deletes ALL existing entries
+  createData()
+  return knex('orders').del()
+    .then(function () {
+      // Inserts seed entries
+      return knex('orders').insert(A)
+    })
+}
